Modify functionality of the line printer tool
The new functionality of the line printer allows you to specify a range
in the form [x-y] where X is the initial line number and Y is the number
of lines from X to print (inclusive). Notice that you can still print a
single line by omitting Y in which case the tool will print the specific
line at X. The script was modified to fail gracefully on invalid line
numbers.