Added tool to follow GitHub accounts via friendship
GitHub offers the ability to follow the activity associated to other
public accounts, they can as well follow yours and in both cases the
username will appear listed in either one of two categories named
followers and following. These pages have a pagination with a maximum
of 51 entries per page (as of August, 2016) distributed in groups of
three per row. Also notice that there is a maximum number of page which
as of today is one hundred.