Added ajax request for the GitHub friendship tool
Using the built-in interface XMLHttpRequest we will send a HTTP POST
request to the internal GitHub API service to create a link between the
two accounts, errors associated with the expiration of the auth token or
any other will be printed in the console. Notice that accounts that are
already linked as friends will still be selected as there is no special
flag that can be used to determine if the account is already in the
friendship list or not, there is a CSS class named "on" but it can be
confused with other elements so we will ignore that.