Added external cookies for GitHub friendship tool
The cookies are loaded from an external INI file that must be located in
the same directory as the script due to limitations in the PhantomJS API
that doesn't allows to retrieve the home directory nor the current user
in the system. We could try to execute the whoami Unix command but it
might not be worth the effort considering the volatility of the session.