Added number of possible accounts in GitHub frienship
The number of accounts listed in the following and follower pages ranges
between one and fifty-two being one of them the username of the main
account in itself so the maximum number of valid accounts per page is as
of today fifty-one. Due to limitations in the HTML DOM we cannot ensure
that all the occurences of the query are actually valid accounts but
this can give a near value.