Fixed WordPress tickets tool with updated patterns
Recently the code for the WordPress forums was updated and the tables
that were being used to display the support tickets changed to be a HTML
unordered list. This commit updates the pattern used to count thenumber
of tickets per page as well as the number of tickets marked as resolved.