Modified password generator to use crypto/rand package
As a password generator the program should not use the math package to
generate the random integers but a cryptographic function like the ones
available in the crypto/rand package. This commit changes this to be a
little bit more secure, although generating random integers to select a
random character in a string is not necessarily secure.