Added IPv6 as a new scanner option in the DNS lookup
Internet Protocol version 6 (IPv6) is the most recent version of the
Internet Protocol (IP), the communications protocol that provides an
identification and location system for computers on networks and routes
traffic across the Internet. IPv6 was developed by the Internet
Engineering Task Force (IETF) to deal with the long-anticipated problem
of IPv4 address exhaustion. IPv6 is intended to replace IPv4.