Install XDebug module in PHP if not avaialble yet
The PHP installation that comes with the Bitnami Nginx stack lacks the
support for debugging, this commit adds the commands to install the
XDebug module via PECL if not available. Notice that the profiler is
enabled by default, the remote debugging is disabled, and the valgrind
dumps are stored in the system's temporary partition.