Added support for Bitnami Nginx stack configuration
The script was targetting the Apache web server installation available
in the common Bitnami stack which comes with the Ruby, {L,M,W}AMP, and
specific components like WordPress. However, the stack composed by an
Nginx installation was being left behind. This commit adds support to
configure the Nginx server the same way the Apache server was being
configured, this includes changes in the owner of the process, the HTTP
and HTTPS port numbers, the location of the document root, and the
addition of new virtual hosts.