Added XDebug command line colors for readability
XDebug (the popular debugger for PHP) comes with most of its settings
disabled by default, this includes the coloring of data types in the
command line which is generally useful, this commit enables this option.
Notice that if you disable it manually and execute the script once again
the option will keep the current value, this is either a one or a zero.