Fix master nameserver during execution of DNS lookups
To stop redundant DNS lookups the user can insert a single dot at the
end of the domain. However, during the execution of the first query to
find one of the nameservers for the subsequent DNS lookups the tool is
getting confused and reporting an error. This commit fixes this issue
and includes a fallback mechanism to use Google's public DNS in case
that no master nameserver is associated to the requested domain.